project/main/requirements.txt
Django	3.1.5
beautifulsoup4	4.9.3
celery	5.0.5
djangorestframework	3.12.2
psycopg2-binary	2.8.6
python-Levenshtein	0.12.0
requests	2.25.1
selenium	3.141.0
